+++ src/Headers.php
@@ -11,6 +11,8 @@ use ArrayIterator;
 use Countable;
 use Iterator;
 use Traversable;
+use Zend\Http\Header\Exception;
+use Zend\Http\Header\GenericHeader;
 use Zend\Loader\PluginClassLocator;
 
 /**
@@ -257,11 +259,11 @@ class Headers implements Countable, Iterator
     public function get($name)
     {
         $key = static::createKey($name);
-        if (! in_array($key, $this->headersKeys)) {
+        if (! $this->has($name)) {
             return false;
         }
 
-        $class = ($this->getPluginClassLoader()->load($key)) ?: 'Zend\Http\Header\GenericHeader';
+        $class = ($this->getPluginClassLoader()->load(str_replace('-', '', $key))) ?: 'Zend\Http\Header\GenericHeader';
 
         if (in_array('Zend\Http\Header\MultipleHeaderInterface', class_implements($class, true))) {
             $headers = [];
@@ -295,7 +297,7 @@ class Headers implements Countable, Iterator
      */
     public function has($name)
     {
-        return (in_array(static::createKey($name), $this->headersKeys));
+        return in_array(static::createKey($name), $this->headersKeys);
     }
 
     /**
@@ -433,17 +435,25 @@ class Headers implements Countable, Iterator
 
     /**
      * @param $index
+     * @param bool $isGeneric
      * @return mixed|void
      */
-    protected function lazyLoadHeader($index)
+    protected function lazyLoadHeader($index, $isGeneric = false)
     {
         $current = $this->headers[$index];
 
         $key = $this->headersKeys[$index];
         /* @var $class Header\HeaderInterface */
-        $class = ($this->getPluginClassLoader()->load($key)) ?: 'Zend\Http\Header\GenericHeader';
+        $class = $this->getPluginClassLoader()->load(str_replace('-', '', $key));
+        if ($isGeneric || ! $class) {
+            $class = GenericHeader::class;
+        }
 
-        $headers = $class::fromString($current['line']);
+        try {
+            $headers = $class::fromString($current['line']);
+        } catch (Exception\InvalidArgumentException $exception) {
+            return $this->lazyLoadHeader($index, true);
+        }
         if (is_array($headers)) {
             $this->headers[$index] = $current = array_shift($headers);
             foreach ($headers as $header) {
@@ -465,6 +475,6 @@ class Headers implements Countable, Iterator
      */
     protected static function createKey($name)
     {
-        return str_replace(['-', '_', ' ', '.'], '', strtolower($name));
+        return str_replace(['_', ' ', '.'], '-', strtolower($name));
     }
 }